Moved delete_all endpoint to Announcements app
