+++ healthcheck/urls.py
@@ -6,5 +6,4 @@ app_name = "healthcheck"
 urlpatterns = [
     path('', health, name='health'),
     path('health/', health, name='health_1'),
-    path('delete/',delete_all,name="delete")
 ]
